Add files via upload
This script describes a fitting mechanism for all the waveforms in the digital optical modules (DOMs) of the IceCube neutrino detectors. For NuTauCC (tau neutrino deep inelastic charged current interactions) events, there are two cascades/showers of Cherenkov radiation from the interaction and decay processes, hence there may be two pulse peaks reflected from the waveform. The curve fitting mechanism is in the form of exponential distribution + Gaussian distribution. This general script looks into all the DOMs in all the events simulated in a single simulation parquet file, and returns two key parameters of these waveforms: the amplitude ratio between the exponential curve and Gaussian curve, and the time separation between the peak of the two curves. Waveforms with comparable amplitudes or with well separated peak times are considered as potential candidates of double cascades waveforms.