Add files via upload
This is a general plotting script for NuTauCC events. There is a general plot on top showing the whole event visualization, including a line indicating the tau neutrino propagating direction, and a cross mark indicating the interaction vertex (where the neutrino interacts with ice molecule). Three subplots below respectively indicate photon depositions from tau minus, from the decay vertex, or from the hadronic vertex (hadronic products from the scattering interaction). The color bar on the right indicates time ranging from the first to the last photon hit in the detector. Each scatter point on the graph represents a DOM in the detector. The size of the scatter point indicates the log of the number of photon hits in this DOM throughout the whole event, and the color of the scatter point indicates the average time at which photons hit the DOMs.