Add files via upload
This is an RNN model built for the classification task. PyTorch RNN is the primary choice because the training samples are waveforms (with amplitude changing with time), which is essentially time series data. RNN has the ability to recognize sequential characteristics of data. The RNN model consists of three layers: one rnn layer and two linear layers. The loss function used is weighted BCELoss function, binary because this is a binary class classification task, and weighted because there are unequal amounts of signal and background DOMs in the training datasets.