Add files via upload
Among all the waveforms (for all the DOMs in all NuTauCC events in a simulation parquet file), a trigger criterion is applied. The c500 trigger is a previous version of trigger used, and it excludes waveforms with cumulative no. of photoelectrons within the 500ns below a certain c500 trigger value.