Add files via upload
Among all the waveforms (for all the DOMs in all NuTauCC events in a simulation parquet file), a trigger criterion is applied. c40 trigger selects trigger events based on the amount of photon deposition in the first 40% of time. This trigger takes into account the factor that not all the waveforms have comparable lengths, and DOMs with shorter waveforms (i.e., shorter time range) may not be the ones that contain less valuable information, so we should not select trigger events based on the cumulative photoelectrons based on a fixed amount of time.