Add files via upload
This script is an example on how to save the trigger events (events that pass through, say, the c40 trigger criterion) into a parquet file such that it can be read easily for future training purposes. The data stored in the parquet file is at waveform-level, i.e., each instance in the parquet file is a waveform, and respectively we can read 5 variables from the waveform: event_id, dom_id, amplitudes, times, and the label of the waveform. DOMs are labelled as 0 or 1, and respectively they are background and signal DOMs.